Move BlockId type to alloy-eip
