+++ crates/eips/src/lib.rs
@@ -22,6 +22,9 @@ extern crate alloc;
 pub mod eip1559;
 pub use eip1559::calc_next_block_base_fee;
 
+pub mod eip1898;
+pub use eip1898::{RpcBlockHash,BlockNumberOrTag,BlockId};
+
 pub mod eip2718;
 
 pub mod eip2930;
